+++ app/services/csv_row_contact_builder.rb
@@ -15,11 +15,17 @@ class CsvRowContactBuilder
   delegate :account_list, to: :import
   delegate :constants, to: CsvImport
 
-  attr_accessor :csv_row, :import, :contact, :person, :spouse
+  attr_accessor :csv_row, :import, :contact, :person, :spouse, :names
+
+  def contact_scope
+    account_list.contacts
+  end
 
   def contact_from_csv_row
     rebuild_csv_row_with_mpdx_headers_and_mpdx_constants
     strip_csv_row_fields
+
+    parse_names
 
     build_contact
     build_addresses
@@ -28,14 +34,29 @@ class CsvRowContactBuilder
     build_email_addresses
     build_phone_numbers
     build_spouse_person
+    build_referral
 
     contact
   end
 
+  def parse_names
+    @names = HumanNameParser.new(csv_row['full_name'] || '').parse
+
+    # first_name and last_name columns take precedence over full_name column
+    @names[:first_name] = csv_row['first_name'] if csv_row['first_name'].present?
+    @names[:last_name] = csv_row['last_name'] if csv_row['last_name'].present?
+    @names[:spouse_first_name] = csv_row['spouse_first_name'] if csv_row['spouse_first_name'].present?
+    @names[:spouse_last_name] = csv_row['spouse_last_name'].presence || @names[:last_name]
+
+    @names[:full_contact_name] = Contact::NameBuilder.new(@names.to_h).name
+
+    @names = @names.with_indifferent_access
+  end
+
   def build_contact
-    self.contact = account_list.contacts.build(
+    self.contact = contact_scope.build(
       church_name: csv_row['church'],
-      name: build_contact_name,
+      name: names['full_contact_name'],
       greeting: csv_row['greeting'],
       envelope_greeting: csv_row['envelope_greeting'],
       status: csv_row['status'],
@@ -48,12 +69,6 @@ class CsvRowContactBuilder
       no_appeals: !true?(csv_row['send_appeals']),
       website: csv_row['website']
     )
-  end
-
-  def build_contact_name
-    first_names = [csv_row['first_name'], csv_row['spouse_first_name']].select(&:present?).to_sentence
-    last_names = [csv_row['last_name'], csv_row['spouse_last_name']].select(&:present?).uniq.to_sentence
-    [last_names, first_names].select(&:present?).join(', ')
   end
 
   def build_addresses
@@ -77,8 +92,8 @@ class CsvRowContactBuilder
   end
 
   def build_primary_person
-    self.person = Person.new(first_name: csv_row['first_name'],
-                             last_name: csv_row['last_name'])
+    self.person = Person.new(first_name: names['first_name'],
+                             last_name: names['last_name'])
     contact.primary_person = person
   end
 
@@ -99,16 +114,29 @@ class CsvRowContactBuilder
   end
 
   def build_spouse_person
-    return if csv_row['spouse_first_name'].blank?
+    return if names['spouse_first_name'].blank?
 
-    spouse = Person.new(first_name: csv_row['spouse_first_name'],
-                        last_name: csv_row['spouse_last_name'].presence || csv_row['last_name'])
+    spouse = Person.new(first_name: names['spouse_first_name'],
+                        last_name: names['spouse_last_name'])
     spouse.email_addresses.build(email: csv_row['spouse_email'],
                                  primary: true) if csv_row['spouse_email'].present?
     spouse.phone_numbers.build(number: csv_row['spouse_phone'],
                                primary: true) if csv_row['spouse_phone'].present?
 
     contact.spouse = spouse
+  end
+
+  def build_referral
+    return if csv_row['referred_by'].blank?
+
+    referrer_contact = Contact::FindFromName.new(contact_scope, csv_row['referred_by']).first
+
+    if referrer_contact.blank?
+      contact.tag_list.add('Missing Csv Referred By')
+      contact.add_to_notes("Referred by: #{csv_row['referred_by']}")
+    else
+      contact.contact_referrals_to_me.build(referred_by: referrer_contact)
+    end
   end
 
   def true?(val)