+++ spec/services/csv_row_contact_builder_spec.rb
@@ -1,8 +1,9 @@
 require 'rails_helper'
 
 describe CsvRowContactBuilder do
-  let(:import) { create(:csv_import_custom_headers, tags: 'csv, test', in_preview: true) }
-  let(:csv_row) { CSV.new(File.open(import.file_path).read, headers: :first_row).first }
+  let!(:import) { create(:csv_import_custom_headers, tags: 'csv, test', in_preview: true) }
+  let!(:csv_row) { CSV.new(File.open(import.file_path).read, headers: :first_row).first }
+  let!(:existing_contact) { create(:contact, name: 'Mary Kim', account_list: import.account_list) }
 
   subject { CsvRowContactBuilder.new(import: import, csv_row: csv_row) }
 
@@ -19,6 +20,7 @@ describe CsvRowContactBuilder do
       'country'              => 'country',
       'email_1'              => 'email_address',
       'first_name'           => 'fname',
+      'full_name'            => 'fullname',
       'greeting'             => 'greeting',
       'envelope_greeting'    => 'mailing_greeting',
       'last_name'            => 'lname',
@@ -27,6 +29,7 @@ describe CsvRowContactBuilder do
       'newsletter'           => 'newsletter',
       'notes'                => 'extra_notes',
       'phone_1'              => 'phone',
+      'referred_by'          => 'referred_by',
       'region'               => 'region',
       'send_appeals'         => 'appeals',
       'spouse_email'         => 'spouse_email_address',
@@ -119,6 +122,78 @@ describe CsvRowContactBuilder do
       expect(spouse.email_addresses.first.email).to eq('jane@example.com')
       expect(spouse.phone_numbers.size).to eq(1)
       expect(spouse.phone_numbers.first.number.in?(['(407) 555-6666', '+14075556666'])).to be(true)
+
+      expect(contact.contact_referrals_to_me.size).to eq(1)
+      contact_referral = contact.contact_referrals_to_me.first
+      expect(contact_referral.referred_by.name).to eq('Mary Kim')
+    end
+
+    context 'referred_by cannot be found' do
+      it 'adds referred_by to contact notes and tag' do
+        Contact.delete_all
+        contact = subject.build
+        expect(contact.contact_referrals_to_me.size).to eq(0)
+        expect(contact.notes).to include('Referred by: Mary Kim')
+        expect(contact.tag_list).to include('missing csv referred by')
+      end
+    end
+  end
+
+  describe 'name parsing' do
+    context 'all name fields are specified' do
+      before do
+        %w(full_name first_name last_name spouse_first_name spouse_last_name).each do |required_key|
+          raise unless import.file_headers_mappings.keys.include?(required_key)
+        end
+      end
+
+      it 'builds names' do
+        contact = subject.build
+        expect(contact.name).to eq('Doe, John and Jane')
+        person = contact.primary_person
+        expect(person.first_name).to eq('John')
+        expect(person.last_name).to eq('Doe')
+        spouse = contact.spouse
+        expect(spouse.first_name).to eq('Jane')
+        expect(spouse.last_name).to eq('Doe')
+      end
+    end
+
+    context 'full_name field is not specified, but other name fields are' do
+      before do
+        import.file_headers_mappings.delete('full_name')
+      end
+
+      it 'builds names' do
+        contact = subject.build
+        expect(contact.name).to eq('Doe, John and Jane')
+        person = contact.primary_person
+        expect(person.first_name).to eq('John')
+        expect(person.last_name).to eq('Doe')
+        spouse = contact.spouse
+        expect(spouse.first_name).to eq('Jane')
+        expect(spouse.last_name).to eq('Doe')
+      end
+    end
+
+    context 'full_name field is specified, but other name fields are not' do
+      before do
+        import.file_headers_mappings.delete('first_name')
+        import.file_headers_mappings.delete('last_name')
+        import.file_headers_mappings.delete('spouse_first_name')
+        import.file_headers_mappings.delete('spouse_last_name')
+      end
+
+      it 'builds names' do
+        contact = subject.build
+        expect(contact.name).to eq('Doey, Johnny and Janey')
+        person = contact.primary_person
+        expect(person.first_name).to eq('Johnny')
+        expect(person.last_name).to eq('Doey')
+        spouse = contact.spouse
+        expect(spouse.first_name).to eq('Janey')
+        expect(spouse.last_name).to eq('Doey')
+      end
     end
   end
 end